tests/wpt/tests/webauthn/getclientcapabilities.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>getClientCapabilities tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
"use strict";

const KNOWN_CAPABILITIES = [
  "conditionalCreate",
  "conditionalGet",
  "hybridTransport",
  "passkeyPlatformAuthenticator",
  "relatedOrigins",
  "signalAllAcceptedCredentials",
  "signalCurrentUserDetails",
  "signalUnknownCredential",
  "userVerifyingPlatformAuthenticator",
];

promise_test(async t => {
  const capabilities = await PublicKeyCredential.getClientCapabilities();
  assert_array_equals(Object.keys(capabilities), KNOWN_CAPABILITIES);
  for (let capability of KNOWN_CAPABILITIES) {
    assert_equals(typeof capabilities[capability], "boolean");
  }
}, "Known capabilities are reported and are sorted");

</script>
